nix/flake-modules/haskell.nix: init
This uses the flake-modules-core module system to implement a
fixed-point for haskell-nix packages. You can now access the haskell.nix
instance of nixpkgs via config.haskell-nix.pkgs in any flake-module,
which means we cannot accidentally import the nixpkgs of haskell.nix
twice. This also allows room for future expansion of this module, which
may come in handy for other reasons.