treewide: update haskell.nix and remove unnecessary follows
We do not need to use inputs.haskell-nix.follows, or specify a
bloated-nixpkgs instance. This code removal produces identical results
as before, except it now defaults to haskell-nix/nixpkgs-unstable. There
are no consequences for this as far as I can tell, so let's just do it
to reduce the amount of code/logic we need to worry about.