treewide: de-obfuscate by removing plutusProjectIn
It seems like this nix code was placed here intentionally to make it
harder to read and modify in the future. The primary problem with
plutusProjectIn is that it messes with the cabal.project for no real
reason, just to be fancy. It creates a fakesrc which is completely
unnecessary, it also prevents us from passing arguments to cabalProject'
for no reason. It is not a function that reduces boiler plate either, as
it doesn't reduce the amount of arguments you need to pass, so I can
only call this obfuscation.