treewide: use default.nix to manage imports of flake-modules
Prior to this, the top-level flake.nix had to be appended to for every
single project. This changes the structure to give each project its own
top-level default.nix which imports other projects. For example,
offchain/default.nix now imports hello-world and hello-world-ui, instead
of the flake.nix doing this.